+++ examples/demo_vdfs_pyplot.jl
@@ -31,7 +31,7 @@ v = zeros(Float64, 3, size(cells)...)
 p = zeros(Float64, size(cells))
 
 vsize = meta.meshes["proton"].vblock_size .* meta.meshes["proton"].vblocks
-vcellids = [Int64[] for _ in cells]
+vcellids = [Int[] for _ in cells]
 vcellf = [Float64[] for _ in cells]
 
 for (i, cell) in enumerate(cells)
